AtCoder/ABC270/C/main_test.go
package main

import (
	"bytes"
	"io"
	"strings"
	"testing"
)

func Test_run(t *testing.T) {
	type args struct {
		stdin io.Reader
	}
	tests := []struct {
		name    string
		args    args
		wantOut string
	}{
		{name: "1", args: args{stdin: strings.NewReader(`5 2 5
		1 2
		1 3
		3 4
		3 5`)}, wantOut: `2 1 3 5`},
		{name: "2", args: args{stdin: strings.NewReader(`6 1 2
		3 1
		2 5
		1 2
		4 1
		2 6`)}, wantOut: `1 2`},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			out := &bytes.Buffer{}
			run(tt.args.stdin, out)
			if gotOut := out.String(); gotOut != tt.wantOut {
				t.Errorf("run() = %v, want %v", gotOut, tt.wantOut)
			}
		})
	}
}
